error handling

+++ src/consensus/constants.py
@@ -53,7 +53,7 @@ class ConsensusConstants:
     MAX_BLOCK_COST_CLVM: int
 
     WEIGHT_PROOF_THRESHOLD: uint8
-    WEIGHT_PROOF_RECENT_BLOCKS: uint8
+    WEIGHT_PROOF_RECENT_BLOCKS: uint32
 
     def replace(self, **changes):
         return dataclasses.replace(self, **changes)
